requirements.txt
django>=1.8.7,<1.9
edx-django-oauth2-provider>=0.3.0,<1.0.0
PyJWT>=1.4.0,<2.0.0
